+++ src/containers/ManageListingsPage/ManageListingsPage.duck.js
@@ -224,7 +224,7 @@ export const queryOwnListings = queryParams => (dispatch, getState, sdk) => {
   dispatch(queryListingsRequest(queryParams));
 
   const { perPage, ...rest } = queryParams;
-  const params = { ...rest, per_page: perPage };
+  const params = { ...rest, perPage };
 
   return sdk.ownListings
     .query(params)